+++ src/components/Breadcrumbs/index.ts
@@ -1,3 +1,3 @@
-export * from './Breadcrumb';
+export * from './Breadcrumbs';
 export * from './BreadcrumbCollapsed';
 export * from './BreadcrumbItem/BreadcrumbItem';
